+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This is the changelog for Tarmac.
 
+## [Unreleased]
+
+## Changed
+
+- Ignore missing script and job metadata
+
 ## 0.1.3
 
 ## Changed
@@ -18,7 +24,7 @@ This is the changelog for Tarmac.
 
 ### Added
 
-- Add CI and release workflows.
+- Add CI and release workflows
 - Add `--base-path` argument to  `tarmac` command
 
 ## 0.1.0